+++ apps/web/src/app/[locale]/landing/RankingSection/GameConsole/GameScreen.tsx
