+++ apps/web/src/app/[locale]/landing/RankingSection/GameConsole/Joystick.tsx
