+++ apps/web/src/app/[locale]/landing/RankingSection/GameConsole/MotionButton.tsx
