+++ apps/web/src/app/[locale]/landing/RankingSection/GameConsole/PixelNoiseEffect.tsx
